# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * COD Order Management API for Algeria
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  pending: "pending",
  confirmed: "confirmed",
  in_transit: "in_transit",
  delivered: "delivered",
  returned: "returned",
  cancelled: "cancelled",
} as const;

export interface Order {
  id: number;
  orderNumber: string;
  status: OrderStatus;
  customerName: string;
  customerPhone: string;
  customerPhone2?: string | null;
  wilaya: string;
  commune: string;
  address: string;
  productName: string;
  productId?: number | null;
  quantity: number;
  unitPrice: number;
  totalAmount: number;
  shippingCost: number;
  notes?: string | null;
  agentId?: number | null;
  agentName?: string | null;
  customerId?: number | null;
  createdAt: string;
  updatedAt: string;
  deliveredAt?: string | null;
}

export interface OrderListResponse {
  orders: Order[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export interface CreateOrderRequest {
  customerName: string;
  customerPhone: string;
  customerPhone2?: string | null;
  wilaya: string;
  commune: string;
  address: string;
  productName: string;
  productId?: number | null;
  quantity: number;
  unitPrice: number;
  shippingCost: number;
  notes?: string | null;
  agentId?: number | null;
  customerId?: number | null;
}

export interface UpdateOrderRequest {
  customerName?: string;
  customerPhone?: string;
  customerPhone2?: string | null;
  wilaya?: string;
  commune?: string;
  address?: string;
  productName?: string;
  productId?: number | null;
  quantity?: number;
  unitPrice?: number;
  shippingCost?: number;
  notes?: string | null;
  agentId?: number | null;
}

export type UpdateStatusRequestStatus =
  (typeof UpdateStatusRequestStatus)[keyof typeof UpdateStatusRequestStatus];

export const UpdateStatusRequestStatus = {
  pending: "pending",
  confirmed: "confirmed",
  in_transit: "in_transit",
  delivered: "delivered",
  returned: "returned",
  cancelled: "cancelled",
} as const;

export interface UpdateStatusRequest {
  status: UpdateStatusRequestStatus;
  notes?: string | null;
}

export interface Customer {
  id: number;
  name: string;
  phone: string;
  phone2?: string | null;
  wilaya: string;
  commune: string;
  address: string;
  totalOrders: number;
  deliveredOrders: number;
  returnedOrders: number;
  totalSpent: number;
  notes?: string | null;
  createdAt: string;
}

export interface CustomerListResponse {
  customers: Customer[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export interface CreateCustomerRequest {
  name: string;
  phone: string;
  phone2?: string | null;
  wilaya: string;
  commune: string;
  address: string;
  notes?: string | null;
}

export type AgentStatus = (typeof AgentStatus)[keyof typeof AgentStatus];

export const AgentStatus = {
  active: "active",
  inactive: "inactive",
} as const;

export interface Agent {
  id: number;
  name: string;
  phone: string;
  wilaya: string;
  status: AgentStatus;
  totalDeliveries: number;
  successRate: number;
  currentOrders: number;
  notes?: string | null;
  createdAt: string;
}

export interface AgentListResponse {
  agents: Agent[];
  total: number;
}

export type CreateAgentRequestStatus =
  (typeof CreateAgentRequestStatus)[keyof typeof CreateAgentRequestStatus];

export const CreateAgentRequestStatus = {
  active: "active",
  inactive: "inactive",
} as const;

export interface CreateAgentRequest {
  name: string;
  phone: string;
  wilaya: string;
  status: CreateAgentRequestStatus;
  notes?: string | null;
}

export interface Product {
  id: number;
  name: string;
  sku?: string | null;
  defaultPrice: number;
  stock?: number | null;
  category?: string | null;
  description?: string | null;
  active: boolean;
  createdAt: string;
}

export interface ProductListResponse {
  products: Product[];
  total: number;
}

export interface CreateProductRequest {
  name: string;
  sku?: string | null;
  defaultPrice: number;
  stock?: number | null;
  category?: string | null;
  description?: string | null;
  active: boolean;
}

export interface AnalyticsSummary {
  totalOrders: number;
  pendingOrders: number;
  confirmedOrders: number;
  inTransitOrders: number;
  deliveredOrders: number;
  returnedOrders: number;
  cancelledOrders: number;
  totalRevenue: number;
  deliveryRate: number;
  returnRate: number;
  avgOrderValue: number;
  todayOrders: number;
  todayRevenue: number;
}

export interface WilayaStat {
  wilaya: string;
  totalOrders: number;
  deliveredOrders: number;
  revenue: number;
}

export interface StatusStat {
  status: string;
  count: number;
  percentage: number;
}

export interface RevenueTrendPoint {
  date: string;
  revenue: number;
  orders: number;
}

export interface Settings {
  companyName: string;
  companyPhone: string;
  defaultShippingCost: number;
  currency: string;
  lowStockThreshold: number;
  autoConfirmOrders: boolean;
}

export type ListOrdersParams = {
  status?: string;
  wilaya?: string;
  agentId?: number;
  search?: string;
  dateFrom?: string;
  dateTo?: string;
  page?: number;
  limit?: number;
};

export type ListCustomersParams = {
  search?: string;
  wilaya?: string;
  page?: number;
  limit?: number;
};

export type ListAgentsParams = {
  search?: string;
  status?: string;
};

export type GetAnalyticsSummaryParams = {
  dateFrom?: string;
  dateTo?: string;
};

export type GetRevenueTrendParams = {
  days?: number;
};