# lib/api-zod/src/generated/types/createAgentRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * COD Order Management API for Algeria
 * OpenAPI spec version: 0.1.0
 */
import type { CreateAgentRequestStatus } from "./createAgentRequestStatus";

export interface CreateAgentRequest {
  name: string;
  phone: string;
  wilaya: string;
  status: CreateAgentRequestStatus;
  notes?: string | null;
}