# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * COD Order Management API for Algeria
 * OpenAPI spec version: 0.1.0
 */
import type { OrderStatus } from "./orderStatus";

export interface Order {
  id: number;
  orderNumber: string;
  status: OrderStatus;
  customerName: string;
  customerPhone: string;
  customerPhone2?: string | null;
  wilaya: string;
  commune: string;
  address: string;
  productName: string;
  productId?: number | null;
  quantity: number;
  unitPrice: number;
  totalAmount: number;
  shippingCost: number;
  notes?: string | null;
  agentId?: number | null;
  agentName?: string | null;
  customerId?: number | null;
  createdAt: string;
  updatedAt: string;
  deliveredAt?: string | null;
}